Load old and new connection flow depending on feature flag
